requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlmodel==0.0.19
psycopg2==2.9.9
alembic==1.13.1
numpy>=1.23.5
opencv-python>=4.1.1
pillow>=10.3.0
torch == 2.2.2
torchvision == 0.17.2
ultralytics>=8.0.232
setuptools>=65.5.1 
easyocr
pyzbar